Update RayQueryModuleRector to handle attribute change
The refactor method has been updated in RayQueryModuleRector for more effective attribute handling. In the method, SQL attribute of the parameter type has been modified if it's RowInterface and has Named attribute. A function changeSqlAttribute is also added to facilitate this change. This new function changes the attribute name to SQL.